requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cryptography==43.0.3
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dotenv-python==0.0.1
dydantic==0.0.8
e2b==1.3.2
e2b-code-interpreter==1.2.0
exa-py==1.9.1
fastapi==0.115.12
frozenlist==1.5.0
gql==3.5.2
graphql-core==3.2.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
iniconfig==2.1.0
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jsonschema_rs==0.29.1
langchain==0.3.21
langchain-anthropic==0.3.10
langchain-community==0.3.20
langchain-core==0.3.49
langchain-exa==0.2.1
langchain-mcp-adapters==0.0.6
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langchainhub==0.1.21
langgraph==0.3.21
langgraph-api==0.0.38
langgraph-checkpoint==2.0.23
langgraph-checkpoint-mongodb==0.1.2
langgraph-cli==0.1.81
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langmem==0.0.17
langsmith==0.3.19
markdown-it-py==3.0.0
marshmallow==3.26.1
mcp==1.4.1
mdurl==0.1.2
motor==3.7.0
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.3
openai==1.69.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
paramiko==3.5.1
pluggy==1.5.0
propcache==0.3.1
protobuf==5.29.4
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.3
PyNaCl==1.5.0
pyperclip==1.9.0
Pysher==1.0.8
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
semver==3.0.4
sentry-sdk==2.24.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.1.3
starlette==0.46.1
structlog==25.2.0
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
trustcall==0.0.38
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websocket-client==1.8.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
